+++ com.unity.visualeffectgraph/CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - auto recompile new vfx
 - better detection of default shader path
 - Bitfield control.
+- Initial Event Name inspector for visual effect asset and component
+- Subgraphs
 - Move HDRP outputs to HDRP package + expose HDRP queue selection
 - Add exposure weight control for HDRP outputs